Resume Checker/backend/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.4.0
python-dotenv==1.0.0
google-generativeai==0.7.2
google-api-python-client==2.171.0
google-auth==2.40.3
requests==2.31.0
PyJWT==2.10.1 